requirements.txt

scapy
numpy
torch
onnxruntime 
onnxruntime-vitisai
netfilterqueue